Provide new API for retrieving PSK parameters to support optional hint.
Motivation:

The actual implementation of the clientPSKCallback and serverPSKCallback
makes the server psk hint mandatory.

Reading RFC4279 it seems that, on the opposite, the server psk hint
should be provided only on application that mandate it.

https://datatracker.ietf.org/doc/html/rfc4279#section-5.2

Modifications:

clientPSKCallback and serverPSKCallback are now deprecated in favor of
clientPSKProvider and serverPSKProvider.
Those closure receive a context containing the information available
for the callee to decide which PSK identity to use.
I also added the maximum length for the PSK to the context.

Result:

This allow client which doesn't expect any server hint to provide a
PSK identity to use the new API and return the right PSK identity.
This allow server which doesn't want to provide an hint to do so.